requirements.txt
aiohttp
python-telegram-bot==22.2
